Remove beta bar.
