+++ src/workbench/SideBar.tsx
@@ -23,12 +23,11 @@ import PythonLogo from "../common/PythonLogo";
 import ApiArea from "../documentation/ApiArea";
 import IdeasArea from "../documentation/IdeasArea";
 import ReferenceArea from "../documentation/ReferenceArea";
-import { flags } from "../flags";
 import ProjectArea from "../project/ProjectArea";
 import { useRouterState } from "../router-hooks";
 import SettingsMenu from "../settings/SettingsMenu";
 import HelpMenu from "./HelpMenu";
-import PreReleaseNotice, { useReleaseDialogState } from "./PreReleaseNotice";
+import { useReleaseDialogState } from "./release-notice-hooks";
 import ReleaseDialogs from "./ReleaseDialogs";
 import SideBarHeader from "./SideBarHeader";
 import SideBarTab from "./SideBarTab";
@@ -200,12 +199,7 @@ const SideBar = ({
           {panes.map((p) => (
             <TabPanel key={p.id} p={0} height="100%">
               <Flex height="100%" direction="column">
-                <ErrorBoundary>
-                  {p.contents}
-                  {!flags.livePreview && (
-                    <PreReleaseNotice onDialogChange={setReleaseDialog} />
-                  )}
-                </ErrorBoundary>
+                <ErrorBoundary>{p.contents}</ErrorBoundary>
               </Flex>
             </TabPanel>
           ))}
